refactor: block fallback options
